// src/app/@core/data/models/datos_pagador/datos_pagador.ts
export class DatosPagador {
    TERPA_SECUENCIA: string;
    TERPA_ANO_PAGO: string;
    TERPA_NATURALEZA: string;
    TERPA_TDO_CODVAR: string;
    TERPA_NRO_DOCUMENTO: string;
    TERPA_DIGITO_CHEQUEO: string;
    TERPA_RAZON_SOCIAL: string;
    TERPA_PRIMER_APELLIDO: string;
    TERPA_SEGUNDO_APELLIDO: string;
    TERPA_PRIMER_NOMBRE: string;
    TERPA_SEGUNDO_NOMBRE: string;
    TERPA_DIRECCION: string;
    TERPA_TELEFONO: string;
    TERPA_EMAIL: string;
    TERPA_ESTADO_REGISTRO: string;
    TERPA_FECHA_REGISTRO: string;
}